// packages/server/src/interfaces/padron/padron.interface.ts
export interface IPadron {
    munCed: string;
    seqCed: string;
    verCed: string;
    nombres: string;
    apellido1: string;
    apellido2: string;
    fechaNac: Date;
    lugarNac: string;
    cedANum: string;
    cedASeri: string;
    cedASexo: string;
    calle: string;
    casa: string;
    edificio: string;
    piso: string;
    apto: string;
    telefono: string;
    codMunicipio: string;
    codCiudad: string;
    codSector: string;
    codPiel: number;
    codSangre: number;
    sexo: string;
    senasPart: any;
    codProf: any;
    codOcup: number;
    codNacion: number;
    estCivil: string;
    npadre: string;
    nmadre: string;
    nconyugue: any;
    cedulaPadre: any;
    cedulaMadre: any;
    cedulaConyugue: any;
    numPasapo: any;
    numReside: any;
    fechaExpe: Date;
    fechaNatu: Date;
    fechaExpiracion: Date;
    categoria: string;
    tipoCausa: string;
    codCausa: any;
    solicitud: any;
    tipoActa: any;
    actaMun: string;
    actaOfic: string;
    actaLibro: string;
    actaFolio: string;
    actaNumero: string;
    actaAno: string;
    tipoLibro: string;
    actaAutomatizada: any;
    estatus: string;
    fechaCapt: Date;
    fechaDecla: Date;
    fechaCancelacion: Date;
    numeroOficio: number;
    fechaRevalidacion: Date;
    numeroOficioRevalidacion: number;
    padreNombres: string;
    padreApellido1: string;
    padreApellido2: string;
    madreNombres: string;
    madreApellido1: string;
    madreApellido2: string;
    conyugueNombres: string;
    conyugueApellido1: string;
    conyugueApellido2: string;
    donante: any;
}