// packages/server/src/interfaces/matriz/matriz-enums.ts
export enum enumTiposMatriz{

    PoliticamenteExpuesto = 'Politicamente Expuesto',
    Extranjero = 'Extranjero',
    PaisesEnListasNegrasOGris = 'Paises En listas negras o Gris',
    ListaOfac = 'Lista Ofac',
    DineroEnEfectivoPorEncimaDelUmbral = 'Dinero en efectivo por encima del umbral',
    InformacionesIncompleta = 'Informaciones incompleta',
    EnRepresentacionDeUnTercero = 'En representacion de un tercero',
    DocumentoNoVerificado = 'Documento no verificado',
    AntecedentesEnDelitosDeLavado = 'Antecedentes en Delitos de Lavado',
    FuentesDeIngresosNoVerificable = 'Fuentes de Ingresos no verificable',
    OcupacionSujetosObligados = 'Ocupacion (Sujetos Obligados)',


}



// <option value="Politicamente Expuesto">Politicamente Expuesto</option>
// <option value="Extranjero">Extranjero</option>
// <option value="Paises En listas negras o Gris">Paises En listas negras o Gris</option>
// <option value="Lista Ofac">Lista Ofac</option>
// <option value="Dinero en efectivo por encima del umbral">Dinero en efectivo por encima del umbral</option>
// <option value="Informaciones incompleta">Informaciones incompleta</option>
// <option value="En representacion de un tercero">En representacion de un tercero</option>
// <option value="Documento no verificado">Documento no verificado</option>
// <option value="Antecedentes en Delitos de Lavado">Antecedentes en Delitos de Lavado</option>
// <option value="Fuentes de Ingresos no verificable">Fuentes de Ingresos no verificable</option>
// <option value="Ocupacion (Sujetos Obligados)">Ocupacion (Sujetos Obligados)</option>